Issue #235 (auto-labels): Implement auto-labels. Sequences with auto-labels returns only its results.
If only one auto-label exist, then result will be result of it expression, otherwise it
will be an array of results of auto-labeled expressions.
Actions are not permitted with auto-labels.

Examples:

```
start1 =   'a'   @'b' 'c';    // return 'b'
start2 =   'a' @b:'b' 'c';    // return 'b', label `b` can be used in semantic predicates
start3 = a:'a' @b:'b' 'c';    // return 'b', labels `a` and `b` can be used in semantic predicates
//start4 = 'a'   @'b' 'c' {}; // forbidden, compilation error
start5 =  @'a'   @'b' 'c';    // return ['a', 'b']
start6 =  @'a' @b:'b' 'c';    // return ['a', 'b'], label `b` can be used in semantic predicates
```